bug fixes apply screen not loading
