+++ web/rainmaker/dev-packages/egov-estate-dev/src/ui-config/screens/specs/estate/applyResourceManimajra/footer.js
@@ -33,8 +33,9 @@ import { getReviewPurchaser, getReviewCourtCase } from "../applyResource/reviewP
 import { WF_ALLOTMENT_OF_SITE } from "../../../../../ui-constants";
 import { downloadAcknowledgementForm,downloadLetter,downloadEmailNotice,downloadNotice,downloadAmountLetter,downloadHousingBoardLetter} from "../../utils";
 import { getFileUrl, getFileUrlFromAPI } from "egov-ui-framework/ui-utils/commons";
-import { getPMDetailsByFileNumber } from "../apply-manimajra";
-
+import {
+  getSearchResults,
+} from "../../../../../ui-utils/commons";
 
 export const DEFAULT_STEP = -1;
 export const PROPERTY_DETAILS_STEP = 0;
@@ -45,6 +46,47 @@ export const PURCHASER_DOCUMENTS_STEP = 4;
 // export const PAYMENT_DETAILS_STEP = 5;
 export const COURT_CASE_DETAILS_STEP = 5;
 export const SUMMARY_STEP = 6;
+
+
+export const getPMDetailsByFileNumber = async (
+  action,
+  state,
+  dispatch,
+  fileNumber
+) => {
+  let queryObject = [
+    {
+      key: "tenantId",
+      value: getTenantId()
+    },
+    { 
+      key: "fileNumber", 
+      value: fileNumber 
+    }
+  ];
+
+  const payload = await getSearchResults(queryObject);
+
+  if (payload) {
+    let properties = payload.Properties;
+    let owners = properties[0].propertyDetails.owners;
+
+    owners = owners.map(item => ({...item, share: (item.share).toString(), ownerDetails: {...item.ownerDetails, isPreviousOwnerRequired: (item.ownerDetails.isPreviousOwnerRequired).toString()}}));
+
+    let currOwners = owners.filter(item => item.ownerDetails.isCurrentOwner == true);
+    let prevOwners = owners.filter(item => item.ownerDetails.isCurrentOwner == false);
+    let areaSqft = (properties[0].propertyDetails.areaSqft).toString();
+
+    properties = [{...properties[0], propertyDetails: {...properties[0].propertyDetails, owners: currOwners, purchaser: prevOwners, areaSqft: areaSqft}}]
+
+    dispatch(
+      prepareFinalObject(
+        "Properties",
+        properties 
+      )
+    )
+  }
+}
 
 export const moveToSuccess = (data, dispatch, type) => {
   const id = get(data, "id");
